requirements.txt
Flask==2.1.1
pycryptodome==3.15.0
requests==2.26.0
gunicorn==20.1.0
